Move dynamic implementation code into main folder, prefix private modules with underscores.
