+++ src/prefab_classes/__init__.py
@@ -19,8 +19,8 @@ __all__ = [
 __version__ = "v0.12.0"
 
 _imports = [
-    MultiFromImport(".dynamic", ["prefab", "attribute", "build_prefab"]),
-    FromImport(".shared", "KW_ONLY"),
+    MultiFromImport("._class_generator", ["prefab", "attribute", "build_prefab"]),
+    FromImport("._shared", "KW_ONLY"),
     MultiFromImport(".funcs", ["is_prefab", "is_prefab_instance"]),
 ]
