--- src/prefab_classes/dynamic/__init__.py
@@ -1,1 +1,0 @@
-from .prefab import prefab, attribute, build_prefab
